xmp: - Deleted unused openssh/preconf/usr/*      - Moved transmission/preconf/bin/* to transmission/preconf/sbin/www/xmproot/.transmissionconfig/      - Moved DCTCS files from flash memory to hdd      - Add Reinstall option when only boot scripts need to be copied

+++ xmp/programs/transmission/install_transmission.sh
@@ -1,10 +1,16 @@
 #!/bin/sh
 
 echo "Transmission install started."
-/opt/bin/ipkg install transmission
-echo "Transmission install done."
-echo ""
-cp -R /sbin/www/xmp/programs/transmission/preconf/* /
-echo "*/5 * * * * root /root/transmission/transmission-queue" >> /opt/var/cron/crontabs/root
+
+if [ ! -f /sbin/www/xmproot/.transmissionconfig/settings.json ]; then
+   /opt/bin/ipkg install transmission
+   cp -R /sbin/www/xmp/programs/transmission/preconf/* /
+   echo "*/5 * * * * root /root/transmission/transmission-queue" >> /opt/var/cron/crontabs/root
+   /bin/chmod 755 /sbin/www/xmproot/.transmissionconfig/transmission-queue
+else
+   cp -R /sbin/www/xmp/programs/transmission/preconf/etc/* /etc
+fi
+
 /bin/chmod 644 /etc/init.d/S227transmission
-/bin/chmod 755 /sbin/www/xmproot/.transmissionconfig/transmission-queue
+
+echo "Transmission install done."
